Move ExpandTestingFixture to new module and test
Move this sphinx directive to a new dedicated module and add some
basic unit tests for it.
