+++ src/globus_sdk/_sphinxext/custom_directives/__init__.py
@@ -1,6 +1,7 @@
 from .add_content_directive import AddContentDirective
 from .automethodlist import AutoMethodList
 from .enumerate_testing_fixtures import EnumerateTestingFixtures
+from .expand_testing_fixture import ExpandTestingFixture
 from .externaldoclink import ExternalDocLink
 from .list_known_scopes import ListKnownScopes
 
@@ -8,6 +9,7 @@ __all__ = (
     "AddContentDirective",
     "AutoMethodList",
     "EnumerateTestingFixtures",
+    "ExpandTestingFixture",
     "ExternalDocLink",
     "ListKnownScopes",
 )
